Second round of refactoring to the garbage collector.
A lot of changes have been made to the garbage collector. Byte-based offsets
and pointers replaced the previous word-based algorythms to make it more
robust and cleaner. C pointers embedded into Java objects are now made using
the placeholder VMPointer class and the type-specific layout has been
entirely removed. The threading system was also modified to reflect the
changes to the way a virtual machine thread was tied to its Java Thread
object. A lot of significant bugs were also fixed in the process.